snapshot manager: filter unavailable dependencies correctly (#20551)
In `refresh --run-program`, we filter out dependencies that become
unavailable due to them being deleted during a refresh. However that's
not the only way a resource can be unavailable during a refresh.

For example a user might create a new resource in the program, and add
that resource as a dependency of a resource that already exists and is
now being refreshed. Because we run the program, the resource being
refreshed now has that dependency, but we're not filtering it out,
because the new resource we're depending on has not been deleted by the
refresh. Instead it's created as a skipped create, which will never end
up in the snapshot.

Fix this by being more aggressive in filtering dependencies, filtering
them every time we have a persisted refresh step. This matches what the
journaling code does, and avoids the snapshot integrity error.

Note that there is a test that catches this
(`TestRefreshWithProviderThatHasDependencies`), but because the
journaling code wasn't copying new resources correctly, it also affected
the state of the resources in the normal snapshot manager, and thus this
issue wasn't caught by the test (and wouldn't have been visible to users
any longer). However we want to fix the journaling code to always copy
resources, and shouldn't rely on that to do the filtering in the first
place.

I noticed this while trying to merge
https://github.com/pulumi/pulumi/pull/20526, but I think this deserves a
separate PR with a separate explanation. No new test, because existing
tests cover this (once #20526 is merged anyway), and I confirmed
manually that the tests pass in combination with that PR.

cc https://github.com/pulumi/pulumi/pull/20541
Closes https://github.com/pulumi/pulumi/pull/20216
Fixes https://github.com/pulumi/pulumi/issues/20215
Fixes https://github.com/pulumi/pulumi/issues/20535