Revert "pass context through to ProjectInfoContext (#20397)" (#20561)
This reverts commit 3e5b321f1a4d6ba9b0e716bbdd7653e812832ad5.

It looks like the daily matrix has been failing consistently starting
18/09 on TestCancelSignal since [the context
PR](https://github.com/pulumi/pulumi/pull/20397) was merged on 17/09
https://github.com/pulumi/pulumi/issues/20469

When an operation is cancelled, the context is canceled, causing
downstream work to cancel or not proceed at all, for example provider
RPC calls for `SignalCancellation` fail immediately because of the
cancelled context.

This also seems to create some test flakiness. The cancellation induces
an error in the program, which is reported as a bail error. This races
with the intended cancellation in the step executor. This can result in
different error reporting for tests that validate cancellation or error
paths.