copy new resources as well as old ones (#20526)
Currently in the journal implementation we only made copies of the old
snapshot, but then added new resources without copying them. This meant
that the engine was still updating pointers, which is not what we want,
as that doesn't translate to the service.

Make the journal make copies of new entries as well to avoid this.

This reveals a few places where we were still relying on the pointers
of resources for correct operation of the journal. This commit fixes
them. In particular:

- It's possible that new resources are later marked as
  Delete/PendingReplacement, so we need to split those fields out into
  DeleteNew/DeleteOld and PendingReplacementNew/PendingReplacementOld.
- RegisterResourceOutputs always piggybacks on a step that has already
  happened, as there is no special step for it. This step has always
  already been executed by the time we get to RegisterResourceOutputs.
  To register the outputs correctly, we therefore need to replace
  the *New()* part of the step with the new outputs, as that's the one
  that's in the intermediate state at this point.
- rebuildDependencies and rebuildBaseState are called at different
  points in the lifecycle.  This can lead to resources that have
  PropertyDependency maps with keys that have an empty list as value
  in the normal snapshot, whereas the journal would have filtered that
  empty list out, because `rebuildDependencies` for the journal runs
  later.  Since empty lists are functionally equivalent to the key not
  being there at all, we can simply account for this in the comparison
  code.