Test for stack config in policy and change Go SDK to support (#20570)
Stacked on top of https://github.com/pulumi/pulumi/pull/20565

This adds a test that stack config is readable from policy packs.

As part of this I've re-jigged the Go SDK to have access to a
`*pulumi.Context` when building the policy pack object. This matches
closely to the current nodejs and python sdks where you can access
things like `pulumi.runtime.isDryRun()` in the root of the file while
building the policy pack.

This required removing the handshake support from the policy pack user
code, but we weren't using that. If we want to add handshake support it
could be an option to `policyx.Main` rather than
`policyx.NewPolicyPack`.