Remove Disabled from policy summary events
We're not currently planning to show which policies were disabled in the Pulumi Cloud UI via these summary events (if Pulumi Cloud needs to, it can look at the policy pack's config to determine which policies are enabled/disabled), so there's no need to send this extra data as part of these events. Remove them for now.