Remove unneeded data from policy summary events (#20589)
Two commits that removes unneeded data from the recently added policy
summary events:

1. **Remove NotApplicable from policy summary events**: This data is
actually quite noisy with the latest policy SDKs, which sends reasons
why each policy is not applicable. We aren't currently planning to show
which policies aren't applicable in the UI, so remove this unnecessary
data from the events for now.

2. **Remove Disabled from policy summary events**: We're not currently
planning to show which policies were disabled in the Pulumi Cloud UI via
these summary events (if Pulumi Cloud needs to, it can look at the
policy pack's config to determine which policies are enabled/disabled),
so there's no need to send this extra data as part of these events.
Remove them for now.