add a version to the journal entry (#20599)
We want to be sure to be able to evolve journal entries. Add a version
to each journal entry, so we can do that. We could just version them by
negotiating the version number with the service, but this makes things
safer.

Based on top of https://github.com/pulumi/pulumi/pull/20559