simplify the journaler interface
We don't actually need separate Begin/End/Write operations in the
journal interface, if we always just pass it journal entries.  The
write code changes a little in that the journal snapshot code already
transforms the write into a journal entry, but that also fixes a bug
where we didn't correctly assign a sequence number to the journal
entry, so I think this is an improvement either way.

This also matches the design for the HTTP API we currently have.