validate that sequence IDs are set, and send error events if not
The Sequence ID field should always be set in all journal entries.
Currently we don't validate that anywhere.  Add validation for that,
and keep the errors somewhere for it.

Note that we don't want to use the `contract.*` functions here,
because we share this code with the service, and want to make sure
that we don't crash the service.  Instead this allows us to send these
errors to the "ErrorEvent" and record them for further debugging.

This also sets up the infrastructure for potential further validation
we might want to do.