Pass opts, name, and type to initialize (#20618)
It's quite common for components to need access to the name, type, or
resource options of the component inside their initialize method.
Currently to do that they have to smuggle this data in with the
arguments, for example see awsx-classic:
https://github.com/pulumi/pulumi-awsx/blob/c04545f2a894e69a8d1b53fa66fb5a7a6f05dddd/awsx-classic/ec2/vpc.ts#L284

This doesn't work well with future direction of wanting to record
component resource inputs to state, where we just want the plain input
arguments being passed down to `ComponentResource` so we can record them
cleanly.

To support cases like awsx we can just pass opts, name, and type to
initialize directly, from the ComponentResource constructor. With them
being passed directly they'll no longer need smuggling in via args.