stop including user auth info in plugin spec name
For plugins that come from Git sources, we use the URL (in a slightly
adjusted way) to name them on disk.  This works until you have a URL
that includes authentication info. In that case the name might end up
being too long, and also we shouldn't write authentication info to
disk like this anyway.

Fix this by removing the authentication info from the URL before using
the URL to generate the name.