Elide function values from resource inputs (#20621)
As part of https://github.com/pulumi/pulumi/issues/10533 we want to send
component inputs to the engine. One component input is the Stack itself,
and for NodeJS that always has an input property "init" which is the
main function to run to make the stack. We don't want to send that, or
any other functions that people happen to be passing down to components,
to the engine.

Currently values of this type fall into the final clause of
`serializeProperty` which tries to treat it as an object and serialise
all the keys, this results in things like `init = {}` being sent to the
engine.

With this change these values are just returned as `undefined` and so
dropped completely from the input struct sent to the engine.