upgrade golangci-lint to the latest version (#20627)
Just keeping golangci-lint up to date. In particular this version uses a
newer version of gofumpt, which wants named return values also to be
named in the return statement. I used `make lint_fix` to fix this up,
and generate the changes in the PR other than the updated versions.