validate that sequence IDs are set, and send error events if not (#20611)
The Sequence ID field should always be set in all journal entries.
Currently we don't validate that anywhere. Add validation for that, and
keep the errors somewhere for it.

Note that we don't want to use the `contract.*` functions here, because
we share this code with the service, and want to make sure that we don't
crash the service. Instead this allows us to send these errors to the
"ErrorEvent" and record them for further debugging.

This also sets up the infrastructure for potential further validation we
might want to do.

/cc @pgavlin. There's actually not too much validation we need to do
here, as we didn't have any `contract.*` validations in the replayer in
the first place. Note that there are still `contract.*` validations in
the snapshotter, that's only run locally (e.g.
https://github.com/pulumi/pulumi/blob/master/pkg/engine/journal_snapshot.go#L210).
We can keep those as we'll never run that code on the service.

That said, if we find any other validations we want to add, it will be
easy to do with this code.

Based on top of https://github.com/pulumi/pulumi/pull/20610