feat: enable `env` for `pulumi about env` in `pulumi about --help` (#20633)
Previously `pulumi about env` was not able to be found when a user ran
`pulumi about --help` this is because we did not have enough environment
variables documented. This has improved significantly, and while it may
not be every environment variable, its enough, to help users find most
things they need.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>