send all diag logs in diff display to stderr (#20632)
We currently send warnings and errors in the diff display to stderr, but
info logs to stdout. We used to do this everywhere, but in
https://github.com/pulumi/pulumi/pull/20461 we decided it would be
better to send all of them to stderr, so the users can filter them as
needed. Also stderr seems like the right place for diagnostics logs, as
they are not the main information the user is looking for, which
`stdout` is meant for, but secondary information, which is what the
`stderr` stream is usually for.

In that PR we didn't change that for the diff display though. Do that
now, to match what we do elsewhere in the CLI, and also to allow users
to filter out diag logs, e.g. for compiler output.

Fixes https://github.com/pulumi/pulumi/issues/20504 (at least one user
confirmed this would be an acceptable solution for the issue for them)
Closes https://github.com/pulumi/pulumi/pull/20582