[backend] Add benchmarks for snapshot managers (#20614)
Add a new set of benchmarks for measuring the performance of our various
snapshot managers (mainly the delta patcher and the in-progress
journaler).

Each benchmark case is an instantation of a general case that runs an
update from an empty base snapshot. The update itself s captured by the
case. Each case is itself parameterizable by the snapshot manager used
for persistence.

There are two kinds of cases: one performs an update by issuing a
set of mutually-independent resources with state that contains random
data, and one replays the registrations and reads implied by the
resources in a state file.

For the former cases, the number of resources and the amount of random
data stored in state are parameterizable. A set list of instantiations
is used to drive the top-level benchmark.

For the latter cases, the benchmarks accept a comma-separated list of
state files in the PULUMI_TEST_SNAPSHOT_REPLAY environment variable.
No state files are included in the test data by default to to the
potential for including sensitive information.

All benchmarks use an httptest server to serve the tiny subset of the
Pulumi Cloud API needed for persistence. This server measures the number
total number of bytes sent over the wire and the total number of calls
made to the server. These metrics are reported by the benchmarks.