replace IDs in lifecycletest golden files (#20626)
lifecycletests are currently quite annoying to re-run under
PULUMI_ACCEPT, because at every rerun all the IDs change, which means
either you check in all the new files, or you go test by test. Either of
those is frustrating to do.

Fix this by introducing stable IDs, and replacing all the testdata using
those. Note that these IDs are only stable as long as tests are not run
in parallel, as events could overlap in such a case. So this doesn't fix
things completely, but I think it's better than what we had before. Not
sure we can fix this problem completely, since some stuff is never going
to be in the same order as it was in previous test runs.

Not entirely sure it's worth the churn, but since I wrote it already I'm
turning it into a PR.