journal: fix snapshot integrity issue when resource is deleted in refresh (#20596)
When a refresh happens, pulumi might delete some resources from the
snapshot. It does so in `rebuildBaseState` right after the refresh. The
journal keeps its own copy of the snapshot (to match what the service
has avaiblable), which doesn't have the same delete step.

This means that after a refresh the indices of resources in the base
snapshot copy might be out of sync with the indices of the resources in
the snapshot the engine is using. In turn that means that when we look
up indices of resources in the engine snapshot by pointer, we may now
get the wrong index of a resource in the base snapshot.

To fix this, we need to do a partial replay of the journal after the
refresh is done. This is safe, because this "rebuiltBaseState" function
runs synchronously in the code, right after the refresh has run, and
before any further operations are happening. We also know that we can
rebuild the base state in the exact same way the engine rebuilds its
state.

We could add the whole snapshot to this step, to avoid needing to do a
rebuild, but that would mean we need to send the whole snapshot at that
point in time to the service, which is expensive and probably not
something we want to do.

(Based on top of https://github.com/pulumi/pulumi/pull/20559, as that
way I won't need to implement parts of this twice, and to avoid
conflicts. In theory it's independent, but stacking it makes things
easier).

---------

Co-authored-by: Pat Gavlin <pat@pulumi.com>