show request ID on service errors
When we get errors from the service, we currently have no way of
correlating them with the logs on the backend. The service already
sets a `X-Pulumi-Request-ID` header for this purpose, but we don't
show it anywhere in the CLI.

To facilitate debugging issues, print out the request ID to the user
if the header is set.