Generate a .gitignore file inside local Node.js SDKS (#20643)
The generated SDKs can optionally be comitted to version control.
Depending on the project setup and package manager used, it is possible
that linking the local SDK into the project will create `node_modules`
inside the local SDK’s folder, but we want’t to avoid adding this to
version control. The same goes for the `bin` folder that contains the
compiled javascript.

Fixes https://github.com/pulumi/pulumi/issues/20630

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>