don't include rootstack when --include-parents is given
We had a wrong condition when calculating the parents that need to be
included in a move.  This caused us to always include the root stack
when --include-parents was given, if the resource to be moved had a
parent. Fix that condition, and update the test to also reflect that.