Support scalar invokes in Go (#20646)
Go fix for https://github.com/pulumi/pulumi/issues/19388

This enables the l2-invoke-scalar conformance test for go. Allowing
functions that have non-object return values to return those types
directly. Similar to how we did it for NodeJS.

I am less confident about this compared to the python implementation,
but this is a start and works for the basic case we're currently testing
in conformance. I do intend to add more scalar invoke tests to
conformance to prove this stuff out better though.