allow resources in same steps to be marked for deletion (#20684)
Since #20061 we generate same steps for untargeted resources during a
destroy. When generating same steps, we assert that `Delete` of the
resource is set to "false". However it is possible for `Delete` to be
set on the resource, and it not being targeted in a destroy in this
case, which triggers the assertion.

We can be a bit more lenient here, and allow same steps to be generated
for resources that are marked for deletion to fix this.

(Found using fuzz testing, the fuzz tester crashed on this)

/cc @Frassle, you might have more context on whether it's okay to have a
resource be marked as `Delete` during a same step. I think it's okay,
but would appreciate a sanity check.