Add support for Python 3.14 (#20433)
Add support for Python 3.14, drops official support for 3.9.

The biggest changes we have to account for, besides dependency upgrades
and some listing fixes, are:

* Asyncio: `asyncio.get_event_loop()` no longer implicitly creates an
event loop if there is none. We now use `asyncio.run(coro)` in entry
points and have to ensure that tests either have a loop as a fixture or
use `pytest.asyncio`.
* Annotations https://peps.python.org/pep-0649/: we need to use
`inspect.get_annotations` as we can no longer rely on `__annotations__`.
`inspect.get_annotations` was added in 3.10.
* Making sure integration tests use the current SDK, not a published
version

Tests against all the supported versions:
https://github.com/pulumi/pulumi/actions/runs/18291273032