Use Language.Link to link python packages
`Language.Link` is changed to take a `map<string, PackageDependency> packages`, where the key is the path to a local SDK to link. This can be a binary SDK, for example a wheel file for Python, or a source directory. This is a breaking change to the protobufs, but the only place where we are currently using Link is in conformance tests.

Link now also returns instructions to provide to the user on how to use the linked SDK, for example `import my_namespace_mypkg as mypkg`. The reason we return this from link is so that the CLI side has no dependency on codegen code.

This first PR moves linking for Python dependencies from the CLI into the Python runtime.

Follow up PRs will update linking for other languages.