Test --local sdks in conformance tests (#20659)
This extends the conformance tests to test "local" sdk generation. This
hits similar code paths to `convert` and is the same sdks that we'd
generate for `package add` (although this doesn't test the `link` part
of that path).

We snapshot the SDKs inside the projects when running in local mode,
this replicates folder layout when running `convert` or `package add`
(i.e. the use of an "sdks" folder).

Currently this works for Python and Go, but has issues for node so it's
currently skipped.