WIP: avoid deserializing deployment from journal manager
When comparing the journal, saving the snapshot, etc. we currently
serialize journal entries, then turn it into a deployment that can be
serialized to JSON, but then unserialize it to get our internal
representation of a snapshot back.  This is because most things
further down the chain expect that to be so.

Unfortunately this causes a bunch more decrypt calls than we currently
have, which costs money. Instead of deserializing the deployment, we
can just serialize the snapshot from the regular snapshot manager
earlier, and can this way compare the two.

The downside here is that it makes the internal implementation a bit
more awkward, with more things operating on the apitype definition.

Note that this is not a problem for the final journal, as the journal
entries will just be sent to the service directly, and thus no decrypt
step is necessary.