Use Language.Link to link python packages (#20612)
`Language.Link` is changed to take a list of `{string: path,
PackageDependency: package}`. The path can be a binary SDK, for example
a wheel file for Python, or a source directory. This is a breaking
change to the protobufs, but the only place where we are currently using
Link is in conformance tests.

Link now also returns instructions to provide to the user on how to use
the linked SDK, for example `import my_namespace_mypkg as mypkg`.

This allows us to push codegen and language specific code out of the CLI
and into the language runtimes. The language runtimes have all the
supporting code to handle package management for their ecosystems, and
we can remove the somewhat shoddy duplicated code used in the `pulumi
package add` command.

This first PR handles linking for Python dependencies. Follow up PRs
will update linking for other languages.