Handle Output forward refs in `unwrap_type`
While adding tests for the helpers in `_types.py` I ran into an issue with `Output` when the forward ref is not resolved.

Eventually we want to unify this code with the helpers in `sdk/python/lib/pulumi/provider/experimental/analyzer.py`. Before that we should increase test coverage. This is a first step.

Follow up from https://github.com/pulumi/pulumi/issues/20607