Add more scalar return tests and fix python maps (#20719)
This adds a few more tests for invokes with scalar returns. Firstly that
secrets work, then that array and map values can be returned.

These don't even compile on Go yet, but they mostly worked on python
except for the map return. This PR fixes sdk-gen that when returning a
map value we don't pick out the one return key, instead just returning
the whole map as is. This seems odd behaviour but it's consistent with
how NodeJS is already working.