don't error on journal errors (#20740)
We're currently running the snapshot journaler in "stealth mode" to
shake out any remaining issues with it. Unfortunately it turns out it
wasn't all that "stealth", as we return errors to the post-step, which
would then make the program fail if there were any bugs in the
journaler.

Allow the combined snapshot manager to only take the errors of some
snapshot managers into consideration when returning errors, so we can
run the journaler more safely for now.

Fixes https://github.com/pulumi/customer-support/issues/2430