Fix scalar maps (#20747)
Noticed this while adding support for scalar returns to Go. There was no
way to return a secret map because we treated maps specially as the root
object so it couldn't be wrapped with the secret marker. This changes
the runtimes and protocol that _all_ scalar returns (including maps) are
a single value nested under an arbitrary (non-private, i.e. don't start
it with "__") key.