make state.Copy() do a deeper copy of the state (#20750)
`state.Copy()` claims to do a deep copy of the state, however for slices
and maps it just assigns them across, meaning the new state will still
operate on the same slice/map after it has been copied.

In particular I noticed this being a problem in the
`NormalizeURNReferences` function, which we [call in the test
journal](https://github.com/pulumi/pulumi/blob/tg%2Fserialized-comparison/pkg/engine/test_journal.go#L186).

Since we're calling `Snap()` of the test journaler for every step, we
call `NormalizeURNReferences` multiple times there. Because we're not
doing a proper deep copy, this means we lose the `Aliases` lists at
every step, since we reset it every time. And then subsequent steps
don't have a access to it anymore, thus creating an invalid snapshot,
because of a dependencies that were aliased, but their URNs not
normalized properly.

We didn't notice this until now, because The regular snapshotter
operates on the same slices, and updated the dependencies lists
correctly beforehand. In https://github.com/pulumi/pulumi/pull/20716 we
however change that and make the regular snapshotter call
`NomalizeURNReferences` on the `apitype` version of the snapshot.

This means we run into the issue above, getting invalid snapshots if
there are aliases and resources that depend on these aliased resources
in later steps.

Fix this by doing a deeper copy of slices and maps. Note that this still
isn't a true deep copy of the state struct, since e.g. created and
modified times are still just regular pointers that aren't deep copied
themselves.