show DestroyV2 failures when parent of provider is updated
When the parent of a provider is updated in the program that `destroy
--run-program` runs, and `--target` is used, we get a snapshot
integrity error.

This happens because we create a replacement for the provider, so it
can be used later, but that replacement stays in the state. As it's
not targeted, it's going to refer to a missing parent.

Unfortunately just removing that parent later doesn't work
either (similarly to what we do in refreshes), haven't quite figured
out why yet.

We can't just same step the provider either similar to what we do for
resources, because the new provider might be needed during the
destroy, e.g. if the provider version was updated.

An attempt to just reset the `Parent` doesn't work either, as we get a
"parent comes after" snapshot integrity issue in that case.

There also seems to be a second issue related to providers here (not
sure it's providers related or otherwise, but I found it this way
anyway), where a resource depending on a provider whose Parent gets
updated seems to be included/targeted even though it's not specified
in the target list, making the step generator fail.

Opening this up as a draft PR mainly in case anyone has any bright
ideas here, as I'm struggling to make progress.