Print replace diffs when warning about import replacements (#20762)
Part of fixing https://github.com/pulumi/pulumi/issues/20486

When a resource wants to replace but has an import ID still set we
warn/error about that, but the message didn't contain any information as
to why it wanted to replace. This adds a list of replace keys and their
old and new values to the error message.