Share CurrentProjectContradictsWorkspace between backends (#20759)
Taking some inspiration from
https://github.com/pulumi/pulumi/pull/19866. One of the few places where
Will had kept things still diverging was the check for if the project
name matched the stack name. But I think we can actually unify this
logic now, and also save the working directory traversal.

Non-backend specific `StackReference`s can be checked for their project
name, we don't need to know what backend were working with to check
that. Also backends keep track of the current project, so there's no
need to go crawl the working directory to find it again.

So this deletes the two private `currentProjectContradictsWorkspace`
functions in the diy and http packages and adds a new public
`CurrentProjectContradictsWorkspace` in the backend package that both
backends then use.

I've also moved the check for the http backend earlier in the apply
function. Up from `createAndStartUpdate` to the method above `apply`.
This matches where the check is done in the diy backend.