Support overriding policy severity on violations (#20767)
This change enables support for specifying a policy severity on
violations on the gRPC protocol. By default, the severity comes from the
policy as returned from `GetAnalyzerInfo`, but if a severity is
specified on a violation diagnostic, the diagnostic severity will be
used.

Part of https://github.com/pulumi/pulumi-policy/issues/425