fuzzing: write out repro if the test times out
It is possible for tests to time out while we're running the fuzz
tester. This is currently very annoying to deal with, as we'll wait
for the overall timeout to time the test out. We also don't currently
write the reproduction test case, so it's very hard to debug.

Similar to snapshot integrity issues, we should never just hang the
program, thus we can always consider this a failure. Fix this by
adding a timeout (20 seconds for now, which should be more enough to
run any lifecycle test, as they are usually not very long tests). Also
write out the repro in case we encounter such a timeout.