implement journaling API and send journal entries to the service directly (#20343)
Implement sending journal entries to the service. This is gated behind
both an environment variable, and the server actually supporting the
feature. For this to work we need server support first, and also setting
`PULUMI_ENABLE_JOURNALING=true`. We still keep the combinedManager
around for error checking as well, just in case.

---------

Co-authored-by: Davide Massarenti <davide@pulumi.com>