refreshV2: avoid hanging forever
In refreshV2, if the we have a resource that's marked pending
replacement in state, and we try to refresh it, we never end up
fulfilling the completion source, making the step generator hang
forever.

The same would be happening if we have a component resource or a
provider that we generate a persisted refresh step for.

Make sure to either fulfill or reject the completion source in all
return paths from `Apply` to avoid this.