only build coverage enabled binaries on linux (#20791)
We only collect coverage data for integration tests on Linux, and
explicitly disable it for
[windows](https://github.com/pulumi/pulumi/blob/master/.github/workflows/ci.yml#L399)
and
[macOS](https://github.com/pulumi/pulumi/blob/master/.github/workflows/ci.yml#L426).

I don't remember the reasoning for this, but I think it was something
about not running on the same platforms during PR checks and in the
merge queue.

Additionally collecting coverage data on Windows is unfortunately often
flaky. Since we're pretty much running the same code on all platforms,
with only very minor differences, the code coverage shouldn't be that
different either way.

However despite not collecting the coverage data on Windows or MacOS, we
still compile the binaries with the `-cover` flag enabled, leading to
test failures because of above mentioned flakyness. Avoid building the
binaries with coverage enabled to reduce the flakyness of our test
suite.

(note this will not reduce the flakyness of everything we consider unit
tests I think, but it should be an improvement nonetheless)