Improve how policy severity is displayed (#20805)
This change improves how severities are displayed for policy violations.
Currently, if a policy has a severity, we'd show it within the
enforcement level's brackets, separated by a colon (e.g. `[mandatory:
high]`). With this change, we move the severity to it's own brackets and
include "severity" so users know what it is (e.g. `[mandatory]
[severity: high]`).

Before: `[mandatory: high]`
 After: `[mandatory] [severity: high]`

Fixes #20804