[python] Run EnsureVenv in RunPlugin
When running a Python plugin, ensure that the venv exists and create it if not. This makes commands like `pulumi package add <path>` work more seamlessly when using local providers.

This will also enable us to make caching of `$PULUMI_HOME/plugins` in CI more robust. The virtualenv folder itslef in a plugin folder should not be cached, because it is not portable across machines. For example if the global Python version changes, a venv will likely break.

Ref https://github.com/pulumi/pulumi/issues/19233
Ref https://github.com/pulumi/pulumi/pull/20755