don't update resource in snapshot during refresh if it's pending replacement
When a resource is pending replacement, we don't need to do anything
in refresh.  We short circuit this in the refresh step here:
https://github.com/pulumi/pulumi/blob/master/pkg/resource/deploy/step.go#L1267.
However in the snapshot managers, we currently still record this step.

For persisted refresh steps, this is incorrect, as this makes the
snapshot managers create a new resource for the refresh, but the
subsequent replacement step still needs to operate with the original
"pending replacement" resource as it's "old" resource. So that
create-replacement step tries to remove the same resource from the
original snapshot, but then adds the resource to the list of new
resources again, after the refresh step already did that.

This was not an issue with the regular refresh code without
`--run-program`, because that *replaced* the old resource in place,
instead of removing the old resource, and then adding a new one to the
list.

By not executing the persistent refresh if we have a resource that is
pending replacement, we get the correct behaviour of only adding the
new resource to the list once, and we don't get a duplicate resource
snapshot error.