fix assert in refreshV2 (#20796)
In https://github.com/pulumi/pulumi/pull/20781, we made it so refreshes
don't hang anymore. However the fix had a problem in that it is now
possible that the completion source returns errors, and would assert in
that case. We want to avoid that assert, so plug the error through and
return it at the appropriate step instead.

Unfortunately I'm not sure how to repro this, but the fuzzer no longer
crashes after this fix.