allow event log to be sent over grpc
The automation API uses the event log to get the summary event from
it. This keeps causing issues, see
e.g. https://github.com/pulumi/pulumi/issues/6768 or
https://github.com/pulumi/pulumi/issues/20788.

This is likely because the library we're using to read these logs is
flaky (see https://github.com/pulumi/pulumi/issues/19888).

It's also not ideal that we need to write out stuff to a file on the
filesystem, and then read it again in the automation API. Instead,
allow sending these events over grpc, which should make them easy to
decode, and avoid these issues.