update tests to explicitly use dotnet 8.0
Update our tests to use dotnet 8.0 as target framework. This makes
these tests work by default on developers machines that have a
supported dotnet framework installed.

In CI this still works because we set `DOTNET_ROLL_FORWARD=Major`, so
it automatically uses newer versions, but we shouldn't expect everyone
to have that set, and it's not obvious from the output that setting
that would work. This makes it a little confusing why this test works
in CI, but not locally, having the same versions of the framework
installed.