refreshv2: only rebuild base state if we have a refresh
In refreshV2, depending on how the program changes, and what resources
are targeted, we might not end up actually refreshing any resources,
but rather just end up with a bunch of same steps.

In particular, this can happen if a resource that's marked as `Delete`
is targeted.

If that happens, we end up incorrectly removing a parent from the
state, because the resource is marked as `Delete`.  The regular
snapshot manager and the journaler handle this correctly, as they only
undangle the parent if there is a refresh step, but we don't handle
this correctly in `rebuildBaseState`.  Don't run `rebuildBaseState` if
there are no refreshes, since it should not be necessary.

This makes sure we get the same result from all the snapshot
implementations in these cases.